+++ frontend/app/src/components/molecules/nav-bar/tenant-switcher.tsx
@@ -72,7 +72,10 @@ export function TenantSwitcher({
                   setOpen(false);
 
                   if (membership.tenant.version === TenantVersion.V1) {
-                    window.location.href = `/tenants/${membership.tenant?.metadata.id}`;
+                    // Hack to wait for next event loop tick so local storage is updated
+                    setTimeout(() => {
+                      window.location.href = `/tenants/${membership.tenant?.metadata.id}`;
+                    }, 0);
                   }
                 }}
                 value={membership.tenant?.slug}