+++ sdks/python/CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to Hatchet's Python SDK will be documented in this changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2025-07-10
+
+### Added
+
+- The `Metrics` client now includes a method to scrape Prometheus metrics from the tenant.
+
+### Changed
+
+- The `Metrics` client's `get_task_metrics` and `get_queue_metrics` now return better-shaped, correctly-fetched data from the API.
+
+## [1.14.4] - 2025-07-09
+
+### Added
+
+- Adds `delete` and `aio_delete` methods to the workflows feature client and the corresponding `Workflow` and `Standalone` classes, allowing for deleting workflows and standalone tasks.
+
+## [1.14.3] - 2025-07-07
+
+### Added
+
+- Adds `remove_null_unicode_character` utility function to remove null unicode characters from data structures.
+
+### Changed
+
+- Task outputs that contain a null unicode character (\u0000) will now throw an exception instead of being serialized.
+- OpenTelemetry instrumentor now correctly reports exceptions raised in tasks to the OTel collector.
+
 ## [1.14.2] - 2025-07-03
 
 ### Added